requirements.txt
PyTDC
rdkit
pandas
scikit-learn
numpy
matplotlib
torch
torchmetrics
